+++ packages/protocol/src/groups/FabricGroupsManager.ts
@@ -4,11 +4,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 import { Fabric } from "#fabric/Fabric.js";
-import { BasicMap, Bytes, Crypto, ImplementationError, InternalError, MatterFlowError, StorageContext } from "#general";
+import { BasicMap, Bytes, Crypto, InternalError, MatterFlowError, StorageContext } from "#general";
 import { GroupKeySet, KeySets, OperationalKeySet } from "#groups/KeySets.js";
-import { PersistedMessageCounter } from "#protocol/MessageCounter.js";
-import { MessageReceptionStateEncryptedWithRollover } from "#protocol/MessageReceptionState.js";
-import { GroupId, NodeId } from "#types";
+import { MessagingState } from "#groups/MessagingState.js";
+import { GroupId } from "#types";
 import { Groups } from "./Groups.js";
 
 export const GROUP_SECURITY_INFO = Bytes.fromString("GroupKey v1.0");
@@ -20,24 +19,15 @@ export class FabricGroupsManager {
     #storage?: StorageContext;
     #fabric: Fabric;
     #groups: Groups;
+    #messagingState: MessagingState;
 
     /** Operationally enhanced variants of the group key sets based on their ID. */
     #keySets = new KeySets();
 
-    /**
-     * Message counter for sending data messages to a group per Operational key. No need to scope to a source node
-     * because we are the sending node
-     * TODO: For management: Make sure to start rotating the key early enough that a former counter-value is not used
-     *  again for the same key.
-     */
-    readonly #groupDataCounters = new Map<string, PersistedMessageCounter>();
-
-    /** Message reception state for data messages per Operational key and source node. */
-    readonly #messageDataReceptionState = new Map<string, Map<NodeId, MessageReceptionStateEncryptedWithRollover>>();
-
     constructor(fabric: Fabric, storage?: StorageContext) {
         this.#fabric = fabric;
         this.#groups = new Groups(fabric, this.#keySets);
+        this.#messagingState = new MessagingState(storage);
 
         // KeySet with ID 0 is always the Fabric IPK, so we initialize from there because this is not stored
         // in Key Management Cluster
@@ -73,6 +63,10 @@ export class FabricGroupsManager {
     /** Operative lookup of the group key sets by their id. */
     get keySets(): KeySets<OperationalKeySet> {
         return this.#keySets;
+    }
+
+    get messaging(): MessagingState {
+        return this.#messagingState;
     }
 
     /** Operative lookup of the group key set id and the key by a group session Id. */
@@ -115,37 +109,6 @@ export class FabricGroupsManager {
         // TODO extend in structure and such
     }
     */
-
-    /**
-     * Return the message counter for sending messages to a group with the given operational key.
-     */
-    messageCounterFor(operationalKey: Uint8Array) {
-        if (!this.#storage) {
-            throw new ImplementationError("Group session cannot be created without storage context.");
-        }
-        const operationalKeyHex = Bytes.toHex(operationalKey);
-        let counter = this.#groupDataCounters.get(operationalKeyHex);
-        if (counter === undefined) {
-            counter = new PersistedMessageCounter(this.#storage, `${operationalKeyHex}-data`);
-            this.#groupDataCounters.set(operationalKeyHex, counter);
-        }
-        return counter;
-    }
-
-    /**
-     * Returns the message reception state for a given source node id and operational key.
-     */
-    receptionStateFor(sourceNodeId: NodeId, operationalKey: Uint8Array) {
-        const operationalKeyHex = Bytes.toHex(operationalKey);
-        let receptionState = this.#messageDataReceptionState.get(operationalKeyHex)?.get(sourceNodeId);
-        if (receptionState === undefined) {
-            receptionState = new MessageReceptionStateEncryptedWithRollover();
-            const keyMap = this.#messageDataReceptionState.get(operationalKeyHex) ?? new Map();
-            keyMap.set(sourceNodeId, receptionState);
-            this.#messageDataReceptionState.set(operationalKeyHex, keyMap);
-        }
-        return receptionState;
-    }
 
     /**
      * Sets new group key set data and pre-calculates all operative data like session ids and operational keys.
@@ -199,12 +162,7 @@ export class FabricGroupsManager {
         // Clean up counters for the group key set
         const operationalKeys = this.#keySets.allKeysForId(groupKeySetId);
         for (const { key } of operationalKeys) {
-            const operationalKeyHex = Bytes.toHex(key);
-            this.#groupDataCounters.delete(operationalKeyHex);
-            if (forDelete) {
-                // If we are deleting the group key set, also delete the persisted counter values
-                await this.#storage?.delete(`${operationalKeyHex}-data`);
-            }
+            await this.#messagingState.removeCounter(key, forDelete);
         }
     }
 }